Move types to svelte files
