+++ examples/api/src/views/App.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { show, hide, setDockVisibility } from '@tauri-apps/api/app'
-  import type { ViewProps } from '../types'
+  import type { ViewProps } from '../App.svelte'
 
   let { onMessage }: ViewProps = $props()
   let dockVisible = $state(true)
